[clean] Remove old compose modal
