+++ src/renderer/components/TimelineSpace.vue
@@ -18,11 +18,10 @@
     <Detail />
   </el-aside>
   <modals></modals>
-  <receive-drop v-show="droppableVisible"></receive-drop>
 </template>
 
 <script lang="ts">
-import { defineComponent, computed, ref, onMounted, onBeforeUnmount } from 'vue'
+import { defineComponent, computed, ref, onMounted } from 'vue'
 import { useRoute } from 'vue-router'
 import { ElMessage } from 'element-plus'
 import { useI18next } from 'vue3-i18next'
@@ -33,28 +32,22 @@ import Compose from './TimelineSpace/Compose.vue'
 import Modals from './TimelineSpace/Modals.vue'
 import Detail from './TimelineSpace/Detail.vue'
 import Mousetrap from 'mousetrap'
-import ReceiveDrop from './TimelineSpace/ReceiveDrop.vue'
 import { AccountLoadError } from '@/errors/load'
 import { TimelineFetchError } from '@/errors/fetch'
-import { NewTootAttachLength } from '@/errors/validations'
-import { EventEmitter } from '@/components/event'
 import { useStore } from '@/store'
 import { ACTION_TYPES } from '@/store/TimelineSpace'
 import { MUTATION_TYPES as GLOBAL_HEADER_MUTATION } from '@/store/GlobalHeader'
 import { MUTATION_TYPES as JUMP_MUTATION } from '@/store/TimelineSpace/Modals/Jump'
-import { ACTION_TYPES as NEW_TOOT_ACTION } from '@/store/TimelineSpace/Modals/NewToot'
 
 export default defineComponent({
   name: 'timeline-space',
-  components: { SideMenu, HeaderMenu, Modals, Contents, ReceiveDrop, Compose, Detail },
+  components: { SideMenu, HeaderMenu, Modals, Contents, Compose, Detail },
   setup() {
     const space = 'TimelineSpace'
     const store = useStore()
     const route = useRoute()
     const i18n = useI18next()
 
-    const dropTarget = ref<any>(null)
-    const droppableVisible = ref<boolean>(false)
     const contentsRef = ref<HTMLElement | null>(null)
 
     const loading = computed(() => store.state.TimelineSpace.loading)
@@ -64,19 +57,10 @@ export default defineComponent({
       await initialize().finally(() => {
         store.commit(`GlobalHeader/${GLOBAL_HEADER_MUTATION.UPDATE_CHANGING}`, false)
       })
-      ;(window as any).addEventListener('dragenter', onDragEnter)
-      ;(window as any).addEventListener('dragleave', onDragLeave)
-      ;(window as any).addEventListener('dragover', onDragOver)
-      ;(window as any).addEventListener('drop', handleDrop)
+
       Mousetrap.bind(['command+t', 'ctrl+t'], () => {
         store.commit(`TimelineSpace/Modals/Jump/${JUMP_MUTATION.CHANGE_MODAL}`, true)
       })
-    })
-    onBeforeUnmount(() => {
-      ;(window as any).removeEventListener('dragenter', onDragEnter)
-      ;(window as any).removeEventListener('dragleave', onDragLeave)
-      ;(window as any).removeEventListener('dragover', onDragOver)
-      ;(window as any).removeEventListener('drop', handleDrop)
     })
 
     const clear = async () => {
@@ -106,56 +90,6 @@ export default defineComponent({
 
       await store.dispatch(`${space}/${ACTION_TYPES.PREPARE_SPACE}`)
     }
-    const handleDrop = (e: DragEvent) => {
-      e.preventDefault()
-      e.stopPropagation()
-      droppableVisible.value = false
-      if (e.dataTransfer?.files.item(0) === null || e.dataTransfer?.files.item(0) === undefined) {
-        return false
-      }
-      const file = e.dataTransfer?.files.item(0)
-      if (file === null || (!file.type.includes('image') && !file.type.includes('video'))) {
-        ElMessage({
-          message: i18n.t('validation.new_toot.attach_image'),
-          type: 'error'
-        })
-        return false
-      }
-      store.dispatch(`TimelineSpace/Modals/NewToot/${NEW_TOOT_ACTION.OPEN_MODAL}`)
-      store
-        .dispatch(`TimelineSpace/Modals/NewToot/${NEW_TOOT_ACTION.UPLOAD_IMAGE}`, file)
-        .then(() => {
-          EventEmitter.emit('image-uploaded')
-        })
-        .catch(err => {
-          if (err instanceof NewTootAttachLength) {
-            ElMessage({
-              message: i18n.t('validation.new_toot.attach_length', { max: 4 }),
-              type: 'error'
-            })
-          } else {
-            ElMessage({
-              message: i18n.t('message.attach_error'),
-              type: 'error'
-            })
-          }
-        })
-      return false
-    }
-    const onDragEnter = (e: DragEvent) => {
-      if (e.dataTransfer && e.dataTransfer.types.indexOf('Files') >= 0) {
-        dropTarget.value = e.target
-        droppableVisible.value = true
-      }
-    }
-    const onDragLeave = (e: DragEvent) => {
-      if (e.target === dropTarget.value) {
-        droppableVisible.value = false
-      }
-    }
-    const onDragOver = (e: DragEvent) => {
-      e.preventDefault()
-    }
 
     const composeResized = (event: { width: number; height: number }) => {
       if (contentsRef.value) {
@@ -165,7 +99,6 @@ export default defineComponent({
 
     return {
       loading,
-      droppableVisible,
       composeResized,
       contentsRef,
       detail
