+++ src/renderer/components/TimelineSpace/Modals.vue
@@ -1,6 +1,5 @@
 <template>
   <div>
-    <new-toot v-if="newTootModal"></new-toot>
     <jump v-if="jumpModal"></jump>
     <image-viewer></image-viewer>
     <list-membership v-if="listMembershipModal"></list-membership>
@@ -14,7 +13,6 @@
 <script lang="ts">
 import { defineComponent, computed } from 'vue'
 import { useStore } from '@/store'
-import NewToot from './Modals/NewToot.vue'
 import Jump from './Modals/Jump.vue'
 import ImageViewer from './Modals/ImageViewer.vue'
 import ListMembership from './Modals/ListMembership.vue'
@@ -26,7 +24,6 @@ import Report from './Modals/Report.vue'
 export default defineComponent({
   name: 'modals',
   components: {
-    NewToot,
     Jump,
     ImageViewer,
     ListMembership,
@@ -37,7 +34,6 @@ export default defineComponent({
   },
   setup() {
     const store = useStore()
-    const newTootModal = computed(() => store.state.TimelineSpace.Modals.NewToot.modalOpen)
     const jumpModal = computed(() => store.state.TimelineSpace.Modals.Jump.modalOpen)
     const reportModal = computed(() => store.state.TimelineSpace.Modals.Report.modalOpen)
     const muteConfirmModal = computed(() => store.state.TimelineSpace.Modals.MuteConfirm.modalOpen)
@@ -45,7 +41,6 @@ export default defineComponent({
     const listMembershipModal = computed(() => store.state.TimelineSpace.Modals.ListMembership.modalOpen)
 
     return {
-      newTootModal,
       jumpModal,
       reportModal,
       muteConfirmModal,
