+++ src/renderer/store/TimelineSpace/Modals.ts
@@ -1,4 +1,3 @@
-import NewToot, { NewTootModuleState } from './Modals/NewToot'
 import ImageViewer, { ImageViewerState } from './Modals/ImageViewer'
 import Jump, { JumpState } from './Modals/Jump'
 import ListMembership, { ListMembershipState } from './Modals/ListMembership'
@@ -17,7 +16,6 @@ type ModalsModule = {
   ImageViewer: ImageViewerState
   ListMembership: ListMembershipState
   MuteConfirm: MuteConfirmState
-  NewToot: NewTootModuleState
   Report: ReportState
   Shortcut: ShortcutState
 }
@@ -29,14 +27,13 @@ const state = (): ModalsState => ({})
 const getters: GetterTree<ModalsState, RootState> = {
   modalOpened: (_state, _getters, rootState) => {
     const imageViewer = rootState.TimelineSpace.Modals.ImageViewer.modalOpen
-    const newToot = rootState.TimelineSpace.Modals.NewToot.modalOpen
     const jump = rootState.TimelineSpace.Modals.Jump.modalOpen
     const listMembership = rootState.TimelineSpace.Modals.ListMembership.modalOpen
     const addListMember = rootState.TimelineSpace.Modals.AddListMember.modalOpen
     const shortcut = rootState.TimelineSpace.Modals.Shortcut.modalOpen
     const muteConfirm = rootState.TimelineSpace.Modals.MuteConfirm.modalOpen
     const report = rootState.TimelineSpace.Modals.Report.modalOpen
-    return imageViewer || newToot || jump || listMembership || addListMember || shortcut || muteConfirm || report
+    return imageViewer || jump || listMembership || addListMember || shortcut || muteConfirm || report
   }
 }
 
@@ -44,7 +41,6 @@ const Modals: Module<ModalsState, RootState> = {
   namespaced: true,
   modules: {
     ImageViewer,
-    NewToot,
     Jump,
     ListMembership,
     AddListMember,
